Refactor search ui, searching by date range
