+++ src/css/components/Launch.css
@@ -15,8 +15,8 @@
 .launch .about { grid-area: about; }
 
 .icon {
-  height: 40px;
-  width: 40px;
+  height: 60px;
+  width: 60px;
   margin-right: 6px;
 
   font-size: 12px;
