+++ time.txt
@@ -1,4 +1,5 @@
- start |   end   | summary
-7:00pm |  7:20pm | Raw result and data displayed
-8:00pm |  8:35pm | Pagination and search
-9:10pm | 10:25pm | Reorganize, styling, exact/fuzzy/regex search types
+  start |   end   | summary
+ 7:00pm |  7:20pm | Raw result and data displayed
+ 8:00pm |  8:35pm | Pagination and search
+ 9:10pm | 10:25pm | Reorganize, styling, exact/fuzzy/regex search types
+10:25pm | 10:55pm | Refactor search ui, searching by date range
